+++ src/main/java/com/transport/management/modules/viaje/ViajeController.java
@@ -1,4 +1,4 @@
-package com.transport.management.controllers;
+package com.transport.management.modules.viaje;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -12,9 +12,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.transport.management.entities.ViajeEntity;
-import com.transport.management.services.ViajeService;
 
 @RestController
 @RequestMapping("/viajes")
