+++ src/main/java/com/transport/management/modules/viaje/ViajeEntity.java
@@ -1,4 +1,4 @@
-package com.transport.management.entities;
+package com.transport.management.modules.viaje;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -7,6 +7,7 @@ import lombok.Setter;
 
 import java.time.LocalDateTime;
 
+import com.transport.management.entities.RutaEntity;
 import com.transport.management.modules.conductor.ConductorEntity;
 import com.transport.management.modules.vehiculo.VehiculoEntity;
 import com.transport.management.utils.abtractBase.BaseEntity;
